Brings in f38cf7606b7 (#1049), which types MONITOR_KEYS so
benchmark_suite_lane.gd parses. Without it the branch's lane script fails to
load ("Cannot infer the type of "key" variable", line 1078, verified with
--check-only), so no lane on this branch -- including open_world_corridor_proof
-- can produce the report the release gate validates, and the Codex request to
capture the proof fixture from the real producer cannot be met. Clean merge,
no conflicts.

Refs #1016

+++ modules/gaussian_splatting/shaders/tile_rasterizer.glsl
@@ -45,7 +45,7 @@ struct Gaussian {
     vec4 rotation;
 
     vec4 sh_dc;
-    float sh_encoded[12];
+    uint sh_encoded[12];  // #1054: raw SNORM10 SH words; uint (not float) so no driver can flush them as denormals
 
     vec3 normal;
